Add support for potential namespace switch, v2.0.0
In preparation for a potential move of the DYK process
to the Wikipedia namespace, the code has been updated
to take the process namespace as a config variable. If
and when a namespace move takes place, changing the
variable should be all that is needed to implement the
move. Since this is a change to the public API, increment
major version number.